requirements.txt
openai>=1.0
pydantic>=2.0
python-dotenv>=1.0
python-telegram-bot>=21.0
psycopg2-binary>=2.9
fastapi>=0.115
uvicorn>=0.30
jinja2>=3.1
python-multipart>=0.0.9
Authlib>=1.3
itsdangerous>=2.2
